% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.package-package.R
\docType{package}
\name{demo.package-package}
\alias{demo.package}
\alias{demo.package-package}
\title{demo.package: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcpri3/demo.package}
  \item Report bugs at \url{https://github.com/mcpri3/demo.package/issues}
}

}
\author{
\strong{Maintainer}: Marie-Caroline Prima \email{marie-caroline.prima@mnhn.fr} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
